Refactor dev_suite modules to use Utils namespace for ErrorHandler, Emoji and Logger
