+++ lib/dev_suite.rb
@@ -2,9 +2,6 @@
 
 module DevSuite
   require_relative "dev_suite/version"
-  require_relative "dev_suite/emoji"
-  require_relative "dev_suite/logger"
-  require_relative "dev_suite/error_handler"
   require_relative "dev_suite/utils"
   require_relative "dev_suite/cli"
   require_relative "dev_suite/performance"
