+++ lib/dev_suite/cli/commands/base.rb
@@ -4,8 +4,6 @@ module DevSuite
   module CLI
     module Commands
       class Base < Thor
-        include ErrorHandler
-
         desc "execute", "Execute the command"
         def execute
           raise NotImplementedError
