+++ lib/dev_suite/cli/commands/version.rb
@@ -5,7 +5,7 @@ module DevSuite
     module Commands
       class Version < Base
         def execute
-          Logger.log("DevSuite version: #{DevSuite::VERSION} is live!", emoji: :start)
+          Utils::Logger.log("DevSuite version: #{DevSuite::VERSION} is live!", emoji: :start)
         end
       end
     end
