+++ lib/dev_suite/cli/main.rb
@@ -25,7 +25,7 @@ module DevSuite
       def execute_command(command_class, *args, **kargs)
         command_class.new.execute(*args, **kargs)
       rescue StandardError => e
-        ErrorHandler.handle_error(e)
+        Utils::ErrorHandler.handle_error(e)
       end
     end
   end
