+++ lib/dev_suite/performance.rb
@@ -8,7 +8,7 @@ module DevSuite
     require_relative "performance/data"
     require_relative "performance/profiler"
     require_relative "performance/profiler_manager"
-    require_relative "performance/reportor"
+    require_relative "performance/reporter"
     require_relative "performance/analyzer"
     require_relative "performance/config"
   end
