+++ lib/dev_suite/performance/reporter/helpers/stat_mappings.rb
@@ -2,7 +2,7 @@
 
 module DevSuite
   module Performance
-    module Reportor
+    module Reporter
       module Helpers
         module StatMappings
           PROFILER_STAT_TITLES = {
