+++ lib/dev_suite/performance/reporter/helpers/table_builder.rb
@@ -2,7 +2,7 @@
 
 module DevSuite
   module Performance
-    module Reportor
+    module Reporter
       module Helpers
         module TableBuilder
           class << self
