+++ lib/dev_suite/utils.rb
@@ -2,6 +2,9 @@
 
 module DevSuite
   module Utils
+    require_relative "utils/emoji"
+    require_relative "utils/logger"
+    require_relative "utils/error_handler"
     require_relative "utils/config_tools"
     require_relative "utils/color"
     require_relative "utils/table"
